get_rate_table

+++ project/bkgd_rxte/get_rate_table.py
@@ -11,9 +11,10 @@ rmffile = 'resp/nicer-rmf6s-teamonly-array50_wo14_34.rmf'
 arffile = 'resp/nicer-consim135p-teamonly-array50_wo14_34.arf'
 energy_bands = {"etot":[0.2,10.0],"e1":[0.2,0.4],"e2":[0.4,0.8],"e3":[0.8,1.6],"e4":[1.6,3.0],"e5":[3.0,6.0],"e6":[6.0,12.0],"e7":[12.0,15.0]}
 outcsv = 'out/bkgd_rxte/bkgd_rxte_rates.csv'
-
+flog = 'out/bkgd_rxte/bkgd_rxte_rates.log'
 record_keywords = ["OBSID","DATEOBS","DATEEND","EXPOSURE","OBJECT","MJDOBS","MJD_DATEOBS","MJD_DATEEND"]
 
+# Init
 dict_out = {}
 for keyword in record_keywords:
 	dict_out[keyword] = []
@@ -25,10 +26,19 @@ for name in energy_bands:
 	keyword = 'rate_%s_%.1fto%.1fkeV_err' % (name,emin,emax)
 	dict_out[keyword] = []
 
+cmd = 'rm -f %s' % outcsv
+print(cmd);os.system(cmd)
+
+# Loop
+f = open(flog,'w')
 for obsid_dir in glob.glob('%s/*' % indir):
 	obsid = os.path.basename(obsid_dir)
 	src_pha = '%s/spec/ni%s_3c50_tot.pi' % (obsid_dir,obsid)
 	bkg_pha = '%s/spec/ni%s_3c50_bkg.pi' % (obsid_dir,obsid)	
+
+	if not os.path.exists(src_pha) or not os.path.exists(bkg_pha):
+		f.write('%s error (file can not be found)' % obsid)
+		continue
 
 	cmd = 'rm -rf tmp_xspec_rate'
 	os.system(cmd)
@@ -53,6 +63,7 @@ for obsid_dir in glob.glob('%s/*' % indir):
 	cmd = 'rm -rf tmp_xspec_rate'
 	os.system(cmd)	
 
+# end 
 for name in energy_bands:
 	emin = energy_bands[name][0]
 	emax = energy_bands[name][1]	
@@ -65,6 +76,7 @@ print(dict_out)
 df = pd.DataFrame(dict_out,columns=record_keywords)
 df.to_csv(outcsv)
 
+f.close()
